+++ middleware/grpc/authentication.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/streamingfast/dauth"
 	"github.com/streamingfast/dauth/middleware"
+	tracing "github.com/streamingfast/sf-tracing"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/metadata"
 )
@@ -27,6 +28,10 @@ func validateAuth(ctx context.Context, path string, authenticator dauth.Authenti
 	md, ok := metadata.FromIncomingContext(ctx)
 	if !ok {
 		md = EmptyMetadata
+	}
+
+	if traceId := tracing.GetTraceID(ctx).String(); traceId != "" {
+		md.Set("x-trace-id", traceId)
 	}
 
 	ctx, err := authenticator.Authenticate(ctx, path, url.Values(md), middleware.RealIP(peerFromContext(ctx), md))