+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+* added `x-trace-id` header in grpc request to dauth
+
 * Added back `secret` plugin support with the form `secret://this-is-the-secret-and-fits-in-the-host-field?[user_id=<value>]&[api_key_id=<value>]`.
 
 * Added continuous authentication support, enable by setting `grpc://localhost:9018?continuous=true`